# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LINUX DO Credit",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/merchant/api-keys": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_key.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/merchant/api-keys/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_key.UpdateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_key.CreateAPIKeyRequest": {
            "type": "object",
            "required": [
                "app_homepage_url",
                "app_name",
                "notify_url"
            ],
            "properties": {
                "app_description": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_homepage_url": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_name": {
                    "type": "string",
                    "maxLength": 20
                },
                "notify_url": {
                    "type": "string",
                    "maxLength": 100
                },
                "redirect_uri": {
                    "type": "string",
                    "maxLength": 100
                },
                "test_mode": {
                    "type": "boolean"
                }
            }
        },
        "api_key.UpdateAPIKeyRequest": {
            "type": "object",
            "properties": {
                "app_description": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_homepage_url": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_name": {
                    "type": "string",
                    "maxLength": 20
                },
                "notify_url": {
                    "type": "string",
                    "maxLength": 100
                },
                "redirect_uri": {
                    "type": "string",
                    "maxLength": 100
                },
                "test_mode": {
                    "type": "boolean"
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}